Refactoring: Cleanup xattrs/metadata package
