+++ pkg/storage/utils/decomposedfs/metadata/errors.go
@@ -16,7 +16,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-package xattrs
+package metadata
 
 import (
 	"os"
