--- pkg/storage/utils/decomposedfs/xattrs/xattrs_suite_test.go
@@ -1,13 +1,0 @@
-package xattrs_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestXattrs(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Xattrs Suite")
-}
